requirements.txt
python-dotenv
chromadb
openai
PyPDF2
streamlit
